backend/pixel_service/requirements.txt
fastapi
uvicorn
redis
sqlalchemy
psycopg2-binary
python-jose
python-dotenv